Refactor word export utilities by consolidating HTML conversion logic into a new htmlConverter module
